Rearranged includes and fixed a lot of warning and added more parameters

+++ registration/src/CPD_Shared.cc
@@ -1,8 +1,12 @@
+#include <math.h>
+
 #include "YgorFilesDirs.h"    //Needed for Does_File_Exist_And_Can_Be_Read(...), etc..
 #include "YgorMisc.h"         //Needed for FUNCINFO, FUNCWARN, FUNCERR macros.
 #include "YgorMath.h"         //Needed for samples_1D.
 #include "YgorString.h"       //Needed for GetFirstRegex(...)
+
 #include "CPD_Shared.h"
+
 #include <chrono>
 using namespace std::chrono;
 
@@ -61,8 +65,8 @@ double GetSimilarity(const Eigen::MatrixXd & xPoints,
     Eigen::MatrixXd tempVector;
 
     double leftSum = 0;
-    for (size_t m = 0; m < mRowsY; ++m) {
-        for (size_t n = 0; n < nRowsX; ++n) {
+    for (int m = 0; m < mRowsY; ++m) {
+        for (int n = 0; n < nRowsX; ++n) {
             tempVector = xPoints.row(n) - AlignedPointSet(yPoints.row(m), rotationMatrix, translation, scale);
             leftSum += postProb(m,n) * tempVector.squaredNorm();
         }
@@ -92,8 +96,8 @@ Eigen::MatrixXd E_Step(const Eigen::MatrixXd & xPoints,
     int dimensionality = yPoints.cols();
 
     auto start = high_resolution_clock::now();
-    for (size_t m = 0; m < mRowsY; ++m) {
-        for (size_t n = 0; n < nRowsX; ++n) {
+    for (int m = 0; m < mRowsY; ++m) {
+        for (int n = 0; n < nRowsX; ++n) {
             tempVector = xPoints.row(n).transpose() - (scale * rotationMatrix * yPoints.row(m).transpose() + t);
             expArg = - 1.0 / (2 * sigmaSquared) * tempVector.squaredNorm();
             expMat(m,n) = exp(expArg);
@@ -101,10 +105,9 @@ Eigen::MatrixXd E_Step(const Eigen::MatrixXd & xPoints,
     }
     auto stop = high_resolution_clock::now();
     auto duration = duration_cast<microseconds>(stop - start); 
-    // std::cout << duration.count() << std::endl; 
 
-    for (size_t m = 0; m < mRowsY; ++m) {
-        for (size_t n = 0; n < nRowsX; ++n) {
+    for (int m = 0; m < mRowsY; ++m) {
+        for (int n = 0; n < nRowsX; ++n) {
             numerator = expMat(m,n);
             denominator = expMat.col(n).sum() + 
                           pow(2 * M_PI * sigmaSquared,((double)(dimensionality/2.0))) * (w/(1-w)) * ((double) mRowsY / nRowsX);
@@ -116,59 +119,6 @@ Eigen::MatrixXd E_Step(const Eigen::MatrixXd & xPoints,
     
     return postProb;
 }
-
-// Eigen::MatrixXd E_Step_Old(const Eigen::MatrixXd & xPoints,
-//             const Eigen::MatrixXd & yPoints,
-//             const Eigen::MatrixXd & BRMatrix,
-//             const Eigen::MatrixXd & t,
-//             double sigmaSquared,
-//             double w) {
-
-//     Eigen::MatrixXd postProb = Eigen::MatrixXd::Zero(yPoints.rows(),xPoints.rows());
-
-//     Eigen::MatrixXd tempVector;
-//     double expArg;
-//     double numerator;
-//     double denomSum;
-//     double denominator;
-
-    // double mRowsY = yPoints.rows();
-    // double nRowsX = xPoints.rows();
-    // double dimensionality = yPoints.cols();
-    
-//     for (size_t m = 0; m < mRowsY; ++m) {
-//         for (size_t n = 0; n < nRowsX; ++n) {
-
-//             auto start = high_resolution_clock::now();
-//             tempVector = xPoints.row(n).transpose() - (BRMatrix * yPoints.row(m).transpose() + t);
-//             expArg = - 1 / (2 * sigmaSquared) * tempVector.squaredNorm();
-//             numerator = exp(expArg);
-            
-//             denomSum = 0;
-//             // auto start = high_resolution_clock::now();
-//             for (size_t k = 0; k < mRowsY; ++k) {
-//                 tempVector = xPoints.row(n).transpose() - (BRMatrix * yPoints.row(k).transpose() + t);
-//                 expArg = - 1 / (2 * sigmaSquared) * tempVector.squaredNorm();
-//                 denomSum += exp(expArg);
-//             }
-//             // auto stop = high_resolution_clock::now();
-//             // auto duration = duration_cast<microseconds>(stop - start); 
-//             // std::cout << duration.count() << std::endl; 
-//             denominator = denomSum + pow(2 * M_PI * sigmaSquared,((double)(dimensionality/2))) * (w/(1-w)) * (double)(mRowsY / nRowsX);
-
-//             postProb(m,n) = numerator / denominator;
-
-//             auto stop = high_resolution_clock::now();
-//             auto duration = duration_cast<microseconds>(stop - start); 
-//             std::cout << duration.count() << std::endl; 
-//         }
-//     }
-
-
-
-//     return postProb;
-
-// }
 
 Eigen::MatrixXd CalculateUx(
             const Eigen::MatrixXd & xPoints,
